Update ColorChanger.py
Allows for editing the layout to be easier since now strings can be entered into the layout dictionary regardless of capitalization. This bug was something I had already caught on the full version, but essentially the previous code would prevent the layout from recognizing a string in the layout dictionary if any character in the string was uppercase.