Update README.md
Important points of clarification in the readme, including the recent fix of inputting strings in the layout, and specifying how Ableton loads a control surface to explain why Ableton has to restart in order to change the color layout. This restart step is avoided in the full version on gumroad.